+++ app/cypress/integration/navigation.spec.js
@@ -16,7 +16,7 @@ context("navigation", () => {
     });
   });
 
-  context("first-level sidebar", () => {
+  context("sidebar", () => {
     it("provides navigation to all main sections of the app", () => {
       cy.get('[data-cy="sidebar"] [data-cy="search"]').click();
       cy.get('[data-cy="page"]').should("contain", "Search");
