+++ app/src/pages/Profile.svelte
@@ -1,6 +1,6 @@
 <script>
   import { Header, Title } from "../DesignSystem";
-  import Layout from "../layouts/NarrowSidebarLayout.svelte";
+  import Layout from "../layouts/SidebarLayout.svelte";
 </script>
 
 <Layout dataCy="page">
