+++ app/src/pages/Projects.svelte
@@ -2,11 +2,11 @@
   import { push } from "svelte-spa-router";
   import { Button, Header, Icon, Title } from "../DesignSystem";
   import Projects from "../components/Projects.svelte";
-  import NarrowSidebarLayout from "../layouts/NarrowSidebarLayout.svelte";
+  import SidebarLayout from "../layouts/SidebarLayout.svelte";
   import { createProject } from "../path.js";
 </script>
 
-<NarrowSidebarLayout dataCy="page">
+<SidebarLayout dataCy="page">
   <Header style="margin-bottom: 24px">
     <div slot="left">
       <Title.Big>My Projects</Title.Big>
@@ -27,4 +27,4 @@
   </Header>
 
   <Projects />
-</NarrowSidebarLayout>
+</SidebarLayout>
